+++ src/main/java/com/baserent/repository/owner/OwnerRepository.java
@@ -5,5 +5,5 @@ import io.micronaut.data.annotation.Repository;
 import io.micronaut.data.repository.CrudRepository;
 
 @Repository
-public interface OwnerRepository extends CrudRepository<Owner, Long> {
+public interface OwnerRepository extends CrudRepository<Owner, Integer> {
 }
